pyap/requirements.txt
Django<=1.11.10
django-ckeditor
django-mptt
pillow
transliterate
gunicorn
sorl-thumbnail
pytz
django-admin-menu
docutils
psycopg2
psycopg2-binary

django-filebrowser
django-filebrowser-no-grappelli
django-simple-search
django-uuslug
django-htmlmin
django-daterange-filter #==1.3.0
django-markdown-deux
django-pagedown
django-simple-math-captcha
